Ms-7/requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.8
botocore==1.34.8
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.9
django-allauth==0.58.2
django-celery-beat==2.5.0
django-crontab==0.7.1
django-decouple==2.1
django-extensions==3.2.3
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.10.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.21.0
jedi==0.19.1
jmespath==1.0.1
kombu==5.3.4
matplotlib-inline==0.1.6
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
